Splict wasm tests into multiple functions and files; use strict mode; use knip
