+++ .gitignore
@@ -43,8 +43,7 @@ test/config.ini
 
 # wasm
 wasm-wrappers/pkg/
-wasm-wrappers/js-bindings-test/test.js
-wasm-wrappers/js-bindings-test/test.js.map
+wasm-wrappers/js-bindings-test/dist/
 
 # 'mintlayer-data' will be mapped to home directories of docker containers, so everything
 # inside it will be generated by the containers.
