+++ wasm-wrappers/js-bindings-test/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "wasm-wrappers-example",
-    "version": "0.1.0",
-    "license": "MIT",
-    "main": "index.js",
-    "type": "module"
-  }
+  "name": "js-bindings-test",
+  "version": "0.1.0",
+  "license": "MIT",
+  "main": "index.js",
+  "type": "module"
+}
