+++ wasm-wrappers/js-bindings-test/web-entry.js
@@ -1,11 +1,11 @@
 import init from "../pkg/wasm_wrappers.js";
 
-import { run_test } from "./test.js";
+import { run_all_tests } from "./dist/main.js";
 
 async function web_run() {
   // Initialize the wasm module
   await init();
-  await run_test();
+  await run_all_tests();
 }
 
 web_run();
